requirements.txt
asgiref==3.4.1
cffi==1.15.0
coverage==6.2
cryptography==36.0.1
Django==3.2.11
django-cors-headers==3.10.1
django-extensions==3.1.5
djangorestframework==3.13.1
flake8==4.0.1
jedi==0.17.0
mccabe==0.6.1
parso==0.8.3
pluggy==1.0.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.3.0
python-jsonrpc-server==0.4.0
python-language-server==0.35.1
pytz==2021.3
sqlparse==0.4.2
ujson==4.3.0
